refactor : fix dto
