--- src/database/dtos/update-user.dto.ts
@@ -1,5 +1,0 @@
-import { PartialType } from "@nestjs/swagger";
-import { UserDto } from "./user.dto";
-
-export class UpadteUserDto extends PartialType(UserDto)  {
-}
