+++ src/modules/user/user.controller.ts
@@ -1,8 +1,8 @@
 import { Body, Controller, Delete, Get, Param, Post, Put } from '@nestjs/common';
 import { UserService } from './user.service';
 import { ApiBody, ApiTags } from '@nestjs/swagger';
-import { UserDto } from '../../database/dtos/user.dto';
-import { UpadteUserDto } from 'src/database/dtos/update-user.dto';
+import { CreateUserDto } from '../../database/dtos/user.dto';
+import { UpadteUserDto } from '../../database/dtos/user.dto';
 
 
 @ApiTags('USERS MGMT')
@@ -12,8 +12,8 @@ export class UserController {
     constructor(private readonly userService: UserService) { }
 
     @Post()
-    @ApiBody({ type: UserDto })
-    create(@Body() payload:UserDto): Promise<UserDto> {
+    @ApiBody({ type: CreateUserDto })
+    create(@Body() payload:CreateUserDto): Promise<CreateUserDto> {
 
         return this.userService.create(payload);
     }
@@ -29,7 +29,7 @@ export class UserController {
     }
 
     @Put(':id')
-    @ApiBody({ type: UserDto })
+    @ApiBody({ type: CreateUserDto })
     update(
         @Body() payload: UpadteUserDto,
         @Param('id') id: string
